Begin of handling notification results

+++ SteakGrillingGuide/Data/DurationSettings.cs
@@ -9,8 +9,8 @@ namespace SteakGrillingGuide.Data
     public class DurationSettings
     {
         public double Thickness { get; set; }
-        public double FirstSide { get; set; }
-        public double SecondSide { get; set; }
-        public double TotalTime => FirstSide + SecondSide;
+        public int FirstSide { get; set; }
+        public int SecondSide { get; set; }
+        public int TotalTime => FirstSide + SecondSide;
     }
 }
